Improves performance of euclidean_distance_field.

On a 512^3 field of ones:
New: 
1.608 MVx/sec (83.475 sec)
1.622 MVx/sec (82.752 sec)
1.672 MVx/sec (80.265 sec)
1.661 MVx/sec (80.814 sec)
1.668 MVx/sec (80.473 sec)
Old: 
1.595 MVx/sec (84.127 sec)
1.638 MVx/sec (81.957 sec)
1.639 MVx/sec (81.897 sec)
1.632 MVx/sec (82.247 sec)
1.622 MVx/sec (82.731 sec)

On a 256^3 field of ones:
New:
2.295 MVx/sec (7.309 sec)
2.371 MVx/sec (7.075 sec)
2.372 MVx/sec (7.072 sec)
2.381 MVx/sec (7.047 sec)
2.362 MVx/sec (7.102 sec)
Old:
2.225 MVx/sec (7.539 sec)
2.213 MVx/sec (7.580 sec)
2.240 MVx/sec (7.491 sec)
2.238 MVx/sec (7.498 sec)
2.238 MVx/sec (7.496 sec)